Add test suffix to migration tests
