+++ owncloudData/src/main/java/com/owncloud/android/data/sharing/shares/datasources/mapper/RemoteShareMapper.kt
@@ -23,6 +23,7 @@ import com.owncloud.android.domain.mappers.RemoteMapper
 import com.owncloud.android.domain.sharing.shares.model.OCShare
 import com.owncloud.android.domain.sharing.shares.model.ShareType
 import com.owncloud.android.lib.resources.shares.RemoteShare
+import com.owncloud.android.lib.resources.shares.ShareType as RemoteShareType
 
 class RemoteShareMapper : RemoteMapper<OCShare, RemoteShare> {
     override fun toModel(remote: RemoteShare?): OCShare? =
@@ -55,7 +56,7 @@ class RemoteShareMapper : RemoteMapper<OCShare, RemoteShare> {
                 sharedWithAdditionalInfo = model.sharedWithAdditionalInfo!!,
                 name = model.name!!,
                 shareLink = model.shareLink!!,
-                shareType = com.owncloud.android.lib.resources.shares.ShareType.fromValue(model.shareType.value),
+                shareType = RemoteShareType.fromValue(model.shareType.value),
                 permissions = model.permissions,
                 sharedDate = model.sharedDate,
                 expirationDate = model.expirationDate,
